+++ src/Stream/Writers/OdsWriter.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FQL\Stream\Writers;
+
+use OpenSpout\Writer\ODS\Writer;
+
+class OdsWriter extends AbstractSpreadsheetWriter
+{
+    protected function createWriter(): Writer
+    {
+        return new Writer();
+    }
+}